Improve ClassFactoryForTestCase cleanup
By default ClassFactoryForTestCase generates its classes in the same package and cleanup this package afterward. 
The problem is that a user can define its own package for the classes to create and those are not cleanup up.

I propose to improve the cleanup to remove those packages if they are empty.