Fix dependency test
We have a dependency test failing because Form has the method #asText that adds a TextAnchor attribute to a string. 
This TextAnchor is in Morphic but in fact it can also be used with forms instead of morphs. 

I'm proposing to move the attribute to Text-Core since it does not necessarily depend on Morphic and I also moved Form>>#asText to Text-Core.

This will fix the dependency test