Improve error on File deletion failure
The file deletion failure uses the #name variable to give info on which file is been deleted. But the method is on the class side, so in reality it returns "File" all the time instead of the file name... On the CI I end up with this kind of failures: `Could not delete file File. Check the file is not open.`. 

I propose to update the error to use the path of the file instead to have a better log and have an easier time to debug things when we have test failures in the CI.